GH-1313 : Moving the test case from GH3513 to GH1313 according to https://github.com/nhibernate/nhibernate-core/pull/3309/files#r1208053155
