+++ src/NHibernate.Test/NHSpecificTest/GH1313/HoldClose.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace NHibernate.Test.NHSpecificTest.GH3513
+namespace NHibernate.Test.NHSpecificTest.GH1313
 {
 	public class HoldClose
 	{
