+++ src/NHibernate.Test/NHSpecificTest/GH1313/Mappings.hbm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <hibernate-mapping xmlns="urn:nhibernate-mapping-2.2" assembly="NHibernate.Test"
-                   namespace="NHibernate.Test.NHSpecificTest.GH3513">
+                   namespace="NHibernate.Test.NHSpecificTest.GH1313">
 
   <class name="HoldClose" >
     <id name="Id" generator="native"/>
